Custom metadata models

+++ src/CommandHandler/DataSpecification/MetadataModel/RemoveNodeCommandHandler.php
@@ -4,7 +4,11 @@ declare(strict_types=1);
 namespace App\CommandHandler\DataSpecification\MetadataModel;
 
 use App\Command\DataSpecification\MetadataModel\RemoveNodeCommand;
+use App\Entity\DataSpecification\MetadataModel\Node\ValueNode;
 use App\Exception\DataSpecification\Common\Model\NodeInUseByTriples;
+use App\Exception\DataSpecification\MetadataModel\NodeHasValues;
+use App\Exception\DataSpecification\MetadataModel\NodeInUseByDisplaySetting;
+use App\Exception\DataSpecification\MetadataModel\NodeInUseByField;
 use App\Exception\NoAccessPermission;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\SecurityBundle\Security;
@@ -13,18 +17,16 @@ use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 #[AsMessageHandler]
 class RemoveNodeCommandHandler
 {
-    private EntityManagerInterface $em;
-    private Security $security;
-
-    public function __construct(EntityManagerInterface $em, Security $security)
+    public function __construct(private EntityManagerInterface $em, private Security $security)
     {
-        $this->em = $em;
-        $this->security = $security;
     }
 
     /**
      * @throws NodeInUseByTriples
      * @throws NoAccessPermission
+     * @throws NodeInUseByField
+     * @throws NodeInUseByDisplaySetting
+     * @throws NodeHasValues
      */
     public function __invoke(RemoveNodeCommand $command): void
     {
@@ -37,6 +39,18 @@ class RemoveNodeCommandHandler
 
         if ($node->hasTriples()) {
             throw new NodeInUseByTriples();
+        }
+
+        if ($node instanceof ValueNode && $node->hasField()) {
+            throw new NodeInUseByField($node->getField()->getTitle());
+        }
+
+        if ($node instanceof ValueNode && $node->hasDisplaySetting()) {
+            throw new NodeInUseByDisplaySetting($node->getDisplaySetting()->getTitle());
+        }
+
+        if ($node instanceof ValueNode && $node->hasValues()) {
+            throw new NodeHasValues();
         }
 
         $this->em->remove($node);
